2.1.3

+++ o2tv/epg.py
@@ -128,19 +128,28 @@ def load_cached_epg():
     open_db(check = 1)
     close_db()
     try:
-      cached_epg_db_url = "http://176.114.248.168:1080/epg_dump.db.gz"
-      cached_epg_db = urlopen(cached_epg_db_url, timeout = 10)
-      content = cached_epg_db.read()
-      with open(addon_userdata_dir + "cached_epg.db.gz", "wb") as f:
-        f.write(content)
-        f.close() 
-      with gzip.open(addon_userdata_dir + "cached_epg.db.gz", "rb") as f:
-        content = f.read()
-        f.close()
-      os.remove(addon_userdata_dir + "cached_epg.db.gz")  
-      with open(addon_userdata_dir + "cached_epg.db", "wb") as f_out:
-        f_out.write(content)
-        f_out.close()
+      if addon.getSetting("uncompressed_cached_epg") == "true":
+        cached_epg_db_url = "http://176.114.248.168:1080/epg_dump.db"
+        cached_epg_db = urlopen(cached_epg_db_url, timeout = 10)
+        content = cached_epg_db.read()
+        with open(addon_userdata_dir + "cached_epg.db", "wb") as f:
+          f.write(content)
+          f.close() 
+      else:
+        cached_epg_db_url = "http://176.114.248.168:1080/epg_dump.db.gz"
+        cached_epg_db = urlopen(cached_epg_db_url, timeout = 10)
+        content = cached_epg_db.read()
+        with open(addon_userdata_dir + "cached_epg.db.gz", "wb") as f:
+          f.write(content)
+          f.close() 
+        with gzip.open(addon_userdata_dir + "cached_epg.db.gz", "rb") as f:
+          content = f.read()
+          f.close()
+        os.remove(addon_userdata_dir + "cached_epg.db.gz")  
+        with open(addon_userdata_dir + "cached_epg.db", "wb") as f_out:
+          f_out.write(content)
+          f_out.close()
+        
       open_db()
       row = None
       for row in db.execute('SELECT version FROM version'):
